Modify challenge answer checking to use hashes
I restored main.js to a previous (unobfuscated) version and modified the
source code to store the answers as SHA-512 hashes instead of plain
text. This prevents someone from opening the source code and seeing all
of the answers.

Of course, someone can still see the answers in the git history ;)

The SubtleCrypto API which provides the SHA-512 hashing function is only
available on a secure connection (HTTPS) and in supporting browsers.
Check the [MDN docs][1] for info on compatibility, but it seems like
most browers have supported this API since around 2014.

There is instructions for adding new questions/answers in the source
code as well.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/digest#browser_compatibility